pyproject.toml
[project]
name = "greens-rain-alert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "W"]